refactor: delete dead code (legacy Manager, unused helpers, no-config constructors)
- Remove legacy manager.Manager + single-impl ProxyManager interface
- Server uses *manager.DBManager directly, drop type-asserts
- Delete unused checker helpers (setters, GetHealthyCount, GroupByStatus, TestSingleProxy)
- Delete dead no-config scraper/checker constructors

~400 net lines removed, no behavior change.
